Code structure changed and new feature
Added option to view sessional ,major GPA without need to input correct query course.

Also re-structured the code in main.js to prevent redudant code and be more module and easier to read for viewers.

Added new paragraph to html for better description.

Adjusted for school year term and that of current to match when seaching for courses